requirements.txt
alembic==1.14.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.6
greenlet==3.1.1
imbalanced-learn==0.13.0
iniconfig==2.1.0
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.6.0
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.2
optuna==4.2.0
packaging==24.2
pandas==2.2.3
pandas-ta==0.3.14b0
pillow==11.1.0
pluggy==1.6.0
psutil==7.0.0
Pygments==2.19.2
pyparsing==3.2.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sklearn-compat==0.1.3
SQLAlchemy==2.0.37
ta==0.11.0
tensorflow==2.19.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
xgboost==2.1.4
schedule==1.2.2
# Note: TA-Lib não pode ser instalado no Windows sem Microsoft Visual C++ Build Tools
# Usando 'ta' e 'pandas-ta' como alternativas funcionais para análise técnica
